corrected stem level data file
Includes the replacing NAs for some of the years (still haven’t figured
out that problem in the database) and corrected DBH values for multiple
stems. There are still some outlier values, but I did not find any way
to confidently correct them. So, I am stuck with some that are negative
growth, and still need to investigate the positive growth outliers.