Add files to repo
The .Rdata file contains data on stems in each of the plots for each
field visit year. The CSV is an export from the Sonoma Access Database
after correcting some of the DBH values in February. I think this
corresponds to correcting the negative values, but not the positive
values, which I still need to finish. I then intend to overwrite or
perhaps just delete this file and have a new one with a more succinct
name.