Update script.
Added read-in of data and initial examination of overstory/understory species data. Created some plots of species abundance: number of plots a species was recorded in, and the total number of a species recorded across the entire plot network. These are split by year, plot establishment (2005) and the most recent sampling season (2014).